backend/pyproject.toml
[tool.poetry]
authors = ["kwiats, radthenone"]
description = "Dashboard project for managing employees and company finances. Django-React project at WSGI"
license = "GNU General Public License v3"
name = "job-schedule"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
argon2-cffi = "^21.3.0"
debugpy = "^1.6.6"
django = "^4.1.6"
django-celery-beat = "^2.4.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
django-extensions = "^3.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.25.1"
flower = "^1.2.0"
hiredis = "^2.0.0"
markdown2 = "^2.4.6"
pillow = "^9.3.0"
python = "^3.10.7"
redis = "^4.3.4"
watchfiles = "^0.18.1"
werkzeug = "^2.2.3"
tablib = "^3.4.0"
django-taggit = "^3.1.0"
django-simple-history = "^3.3.0"
django-model-utils = "^4.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
coverage = "^7.1.0"
django-debug-toolbar = "^3.7.0"
django-silk = "^5.0.3"
django-stubs = "^1.14.0"
dlint = "^0.13.0"
factory-boy = "^3.2.1"
faker = "^16.6.1"
flake8 = "^5.0.4"
flake8-absolute-import = "^1.0.0.1"
flake8-bugbear = "^23.1.20"
flake8-class-attributes-order = "^0.1.3"
flake8-comprehensions = "^3.10.1"
flake8-debugger = "^4.1.2"
flake8-django = "^1.1.5"
flake8-newspaper-style = "^0.4.3"
flake8-no-implicit-concat = "^0.3.4"
flake8-no-pep420 = "^2.3.0"
flake8-print = "^5.0.0"
flake8-pytest-style = "^1.6.0"
flake8-simplify = "^0.19.3"
flake8-warnings = "^0.4.0"
ipython = "^8.6.0"
isort = "^5.12.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
tox = "^4.4.4"

[tool.poetry.group.production]
optional = true

[tool.poetry.group.production.dependencies]
psycopg2-binary = "^2.9.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
